requirements.txt
psycopg2-binary
fastapi
passlib[bcrypt]
pytest
httpx
bcrypt
python-jose
python-dotenv